refactor(schedulers): split engine responsibilities
